backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
pydantic[email]==2.5.0
python-decouple==3.8
aiofiles==23.2.1
Pillow==10.1.0
fastapi-cors==0.0.6
PyPDF2==3.0.1
python-docx==1.2.0
python-magic==0.4.27
pyexiftool==0.4.11
pypdfium2
opencv-python
PyMuPDF
# PostgreSQL driver
psycopg2-binary==2.9.9
